chore: 로그아웃 | 회원탈퇴 위치 변경
프로필 편집 탭 -> 마이페이지 메인 탭

resolves #35
