+++ src/screens/Home/MypageMainScreen.js
@@ -1,6 +1,6 @@
 import React, {useState} from 'react';
 import {ScreenContainer, ScreenHeader} from '../../components';
-import {Text, Image} from 'react-native';
+import {Text, Image, TouchableOpacity} from 'react-native';
 import DonationDialogModal from '../../modal/modals/Home/DonationDialogModal';
 import Profile from '../../components/home/MyPage/Profile';
 import UserInfo from '../../components/home/MyPage/UserInfo';
@@ -10,6 +10,9 @@ import SvgIcon from '../../components';
 import styled from 'styled-components';
 import {SizedBox} from '../../components';
 import LogoutModal from '../../modal/modals/Home/LogoutModal';
+import {useLogoutModalState} from '../../modal/recoil/useModals';
+import font, {Subhead2} from '../../styles/font';
+import {theme, palette} from '../../styles/theme';
 export default function MypageMainScreen({navigation}) {
   const [userInfo, setUserInfo] = useState({
     userId: 2,
@@ -18,6 +21,11 @@ export default function MypageMainScreen({navigation}) {
     profileUrl: null,
     fruitBox: 0,
   });
+  const {openModal} = useLogoutModalState();
+
+  const _unregister = () => {
+    //회원탈퇴 요청
+  };
 
   return (
     <>
@@ -38,9 +46,25 @@ export default function MypageMainScreen({navigation}) {
         <DonationBox n={userInfo.fruitBox} />
         <SizedBox height={24} />
         <Menu navigation={navigation} />
+        <Right>
+          <TouchableOpacity onPress={openModal}>
+            <Subhead2 color={theme.colors.main}>로그아웃 </Subhead2>
+          </TouchableOpacity>
+          <Subhead2 color={palette.gray200}> | </Subhead2>
+          <TouchableOpacity onPress={_unregister}>
+            <Subhead2 color={palette.gray300}>회원탈퇴 </Subhead2>
+          </TouchableOpacity>
+        </Right>
       </ScreenContainer>
       <LogoutModal />
       <DonationDialogModal />
     </>
   );
 }
+const Right = styled.View`
+  display: flex;
+  flex-direction: row;
+  justify-content: flex-end;
+  align-items: center;
+  padding-top: 20px;
+`;
